+++ test/libsolidity/optimizedIRCaching/bytecode_dependency_creation_and_runtime_shared.sol
@@ -1,0 +1,11 @@
+contract A {}
+
+contract C {
+    A a = new A();
+
+    function f() public returns (A) {
+        return new A();
+    }
+}
+// ----
+// cachedObjects: 4