+++ test/libsolidity/optimizedIRCaching/bytecode_dependency_shared_by_multiple_contracts.sol
@@ -1,0 +1,19 @@
+contract A {}
+
+contract C {
+    A a = new A();
+
+    function f() public returns (A) {
+        return new A();
+    }
+}
+
+contract D {
+    A a = new A();
+
+    function f() public returns (A) {
+        return new A();
+    }
+}
+// ----
+// cachedObjects: 6